# defaultPoll.Wait: add poll mechanism to bring some performance improvement