It seems this modification could bring about 4~6% performance improvement under scenario 3. I didn't test other scenarios yet.

```
bash scripts/benchmark_server.sh 3
```

Benchmark of Vanilla:

```
Info: [netpoll] start ..., concurrent: 100, total: 5000000
Info: took 12099 ms for 5000000 requests
Info: requests total: 5000000, failed: 0
Info: [netpoll]: TPS: 413223.68, TP99: 0.81ms, TP999: 1.47ms (b=1024 Byte, c=100, n=5000000)
[NETPOLL@Server] CPU Usage: MIN: 280.28%, TP50: 286.00%, TP90: 292.29%, TP99: 293.29%, MAX: 302.30%, AVG:288.28%
[NETPOLL@Server] Mem Usage: AVG: 15931 KB, MAX: 17508 KB
[NETPOLL@Client] CPU Usage: MIN: 167.00%, TP50: 493.00%, TP90: 506.00%, TP99: 507.51%, MAX: 516.00%, AVG:466.96%
[NETPOLL@Client] Mem Usage: AVG: 86469 KB, MAX: 103240 KB

Info: [netpoll] start ..., concurrent: 500, total: 5000000
Info: took 9739 ms for 5000000 requests
Info: requests total: 5000000, failed: 0
Info: [netpoll]: TPS: 513393.58, TP99: 3.79ms, TP999: 5.44ms (b=1024 Byte, c=500, n=5000000)
[NETPOLL@Server] CPU Usage: MIN: 291.00%, TP50: 300.30%, TP90: 304.00%, TP99: 304.00%, MAX: 307.31%, AVG:299.90%
[NETPOLL@Server] Mem Usage: AVG: 27388 KB, MAX: 30204 KB
[NETPOLL@Client] CPU Usage: MIN: 114.00%, TP50: 577.58%, TP90: 604.60%, TP99: 604.60%, MAX: 621.00%, AVG:529.08%
[NETPOLL@Client] Mem Usage: AVG: 76062 KB, MAX: 105396 KB
```

Benchmark of sleep-poll with 128 max-events:

```
Info: [netpoll] start ..., concurrent: 100, total: 5000000
Info: took 11632 ms for 5000000 requests
Info: requests total: 5000000, failed: 0
Info: [netpoll]: TPS: 429830.47, TP99: 0.75ms, TP999: 1.32ms (b=1024 Byte, c=100, n=5000000)
[NETPOLL@Server] CPU Usage: MIN: 300.00%, TP50: 311.31%, TP90: 316.32%, TP99: 316.32%, MAX: 324.00%, AVG:310.66%
[NETPOLL@Server] Mem Usage: AVG: 17792 KB, MAX: 18840 KB
[NETPOLL@Client] CPU Usage: MIN: 148.15%, TP50: 499.50%, TP90: 507.00%, TP99: 510.51%, MAX: 521.52%, AVG:461.18%
[NETPOLL@Client] Mem Usage: AVG: 104113 KB, MAX: 185636 KB

Info: [netpoll] start ..., concurrent: 500, total: 5000000
Info: took 9111 ms for 5000000 requests
Info: requests total: 5000000, failed: 0
Info: [netpoll]: TPS: 548752.37, TP99: 3.39ms, TP999: 4.61ms (b=1024 Byte, c=500, n=5000000)
[NETPOLL@Server] CPU Usage: MIN: 301.00%, TP50: 313.00%, TP90: 317.00%, TP99: 317.00%, MAX: 321.00%, AVG:313.12%
[NETPOLL@Server] Mem Usage: AVG: 29033 KB, MAX: 31624 KB
[NETPOLL@Client] CPU Usage: MIN: 174.17%, TP50: 606.00%, TP90: 614.61%, TP99: 614.61%, MAX: 623.00%, AVG:559.84%
[NETPOLL@Client] Mem Usage: AVG: 78511 KB, MAX: 109780 KB
```