@Hchenn 2022-02-07
It is not acceptable to sleep in the poller for the time being, which will reduce the sensitivity of the poller, and cannot convince all scenarios. It is more meant to optimize performance for benchmark self.